Add files via upload
## Description [EN]
Car theft script for ESX FiveM servers. Players can steal a car which they will have to deliver to a location to get some dirty money.
Note: This is NOT a job. This is an illegal activity anyone can do.

##
for more detailed information I suggest you open the readme file